refactor(src-types): Move DaoStore types to types file
